+++ src/app/routes/ttts/transactions/placeOrder.ts
@@ -11,13 +11,13 @@ const placeOrderTransactionsRouter = Router();
 import permitScopes from '../../../middlewares/permitScopes';
 import validator from '../../../middlewares/validator';
 
-const chevreAuthClient = new cinerino.chevre.auth.ClientCredentials({
-    domain: <string>process.env.CHEVRE_AUTHORIZE_SERVER_DOMAIN,
-    clientId: <string>process.env.CHEVRE_CLIENT_ID,
-    clientSecret: <string>process.env.CHEVRE_CLIENT_SECRET,
-    scopes: [],
-    state: ''
-});
+// const chevreAuthClient = new cinerino.chevre.auth.ClientCredentials({
+//     domain: <string>process.env.CHEVRE_AUTHORIZE_SERVER_DOMAIN,
+//     clientId: <string>process.env.CHEVRE_CLIENT_ID,
+//     clientSecret: <string>process.env.CHEVRE_CLIENT_SECRET,
+//     scopes: [],
+//     state: ''
+// });
 
 export interface IAcceptedOffer4ttts {
     /**
@@ -49,7 +49,7 @@ placeOrderTransactionsRouter.post(
             // チケットオファー検索
             const eventService = new cinerino.chevre.service.Event({
                 endpoint: cinerino.credentials.chevre.endpoint,
-                auth: chevreAuthClient,
+                auth: req.chevreAuthClient,
                 project: { id: req.project.id }
             });
             const ticketOffers = await eventService.searchTicketOffers({ id: eventId });
@@ -86,6 +86,7 @@ placeOrderTransactionsRouter.post(
                 }
             })({
                 action: new cinerino.repository.Action(mongoose.connection),
+                event: eventService,
                 transaction: new cinerino.repository.Transaction(mongoose.connection)
             });